moved the Player class to a dedicated component
