+++ shooter/index.js
@@ -1,6 +1,7 @@
 import { Screen } from "./modules/Screen.js";
 import { Vector2, InfoDict, Time } from "./modules/System.js";
 import { Keyboard, Mouse } from "./modules/Controllers.js";
+import { Player } from "./modules/Components.js"
 
 const $ = document
 const DEBUG = true
@@ -9,49 +10,11 @@ const putInfo = ((el) => (it) => {
   return el.innerText = it
 })($.querySelector("div#info-pane"))
 
-
-class Player {
-  constructor(coord, size, color, velocity) {
-    this.coord = coord
-    this.size = size
-    this.color = color
-    this.velocity = velocity
-  }
-
-  Update(data) {
-    const keys = data.keys
-    if (keys.up) {
-      this.coord.y -= this.velocity.y
-    } else if (keys.down) {
-      this.coord.y += this.velocity.y
-    }
-
-    if (keys.right) {
-      this.coord.x += this.velocity.x
-    } else if (keys.left) {
-      this.coord.x -= this.velocity.x
-    }
-  }
-
-  Draw(screen) {
-    const ctx = screen.ctx
-    ctx.save()
-    ctx.beginPath()
-    ctx.fillStyle = this.color
-    ctx.shadowBlur = 6
-    ctx.shadowColor = "#af1313"
-    ctx.arc(this.coord.x, this.coord.y, this.size, 0, Math.PI * 2)
-    ctx.fill()
-    ctx.closePath()
-    ctx.restore()
-  }
-}
-
 const screen = new Screen($.querySelector("canvas#main"))
 screen.Resize(innerWidth, innerHeight)
 screen.SetBackground("black")
 
-const p = new Player(screen.origin, 25, "#a80707", new Vector2(4, 4))
+const p = new Player(screen.origin, 25, "#a80707", new Vector2(6, 6))
 screen.AddComponent(p)
