+++ shooter/modules/Components.js
@@ -1,0 +1,1 @@
+export * from "./components/Player.js"
